Add handsontable manager into build proccess Refactoring handsontable manager, remove duplications Describe handsontable manager module ESLint fix
