+++ docs/.vuepress/enhanceApp.js
@@ -1,4 +1,6 @@
-// import RouterLink from './theme/components/RouterLink.vue';
+const { applyToWindow, instanceRegister } = require('./handsontable-manager');
+
+applyToWindow();
 
 const buildRegisterCleaner = register => (to, from) => {
   if (to.path === from.path) {
@@ -7,6 +9,7 @@ const buildRegisterCleaner = register => (to, from) => {
   if (register === undefined) {
     // eslint-disable-next-line
     console.warn('The register doesn\'t exists');
+
     return;
   }
   register.destroyAll();
@@ -22,6 +25,7 @@ const buildActiveHeaderLinkHandler = () => {
       }
       // eslint-disable-next-line
       activeLink = document.querySelector(`.table-of-contents a[href="${to.hash}"]`);
+
       if (activeLink) {
         activeLink.classList.add('active');
       }
@@ -31,8 +35,8 @@ const buildActiveHeaderLinkHandler = () => {
 
 export default ({ router, isServer }) => {
   if (!isServer) {
-    // eslint-disable-next-line
-    router.afterEach(buildRegisterCleaner(handsontableInstancesRegister));
+
+    router.afterEach(buildRegisterCleaner(instanceRegister));
     router.afterEach(buildActiveHeaderLinkHandler());
   }
 };
