+++ docs/next/guides/cell-features/autofill-data.md
@@ -7,11 +7,11 @@ tags:
   - smart fill
 ---
 
-# AutoFill data
+# Autofill data
 
 [[toc]]
 
-## Auto fill in all directions
+## Autofill in all directions
 
 Notice the little square (fill handle) in the corner of the selected cell. You can drag it (drag-down) to repeat the values from the cell. Double click the fill handle in cell B4 (value "30") to fill the selection down to the last value in neighbouring column, just like it would in Excel or Google Sheets.
 
@@ -39,7 +39,7 @@ hot1.loadData(data);
 ```
 :::
 
-## Auto fill in vertical direction only with creating new rows
+## Autofill in vertical direction only with creating new rows
 
 In this configuration fill handle is restricted to move only vertically. Additionally, to adding new rows when it comes to the edge of the table bottom we need change **autoInsertRow** to **true**.
